TimelineList using logic hook
