+++ SharePointFramework/ProjectWebParts/src/components/ProjectTimeline/TimelineList/ITimelineListProps.tsx
@@ -1,0 +1,6 @@
+import { IBaseWebPartComponentProps } from 'components/BaseWebPartComponent';
+
+
+export interface ITimelineListProps extends IBaseWebPartComponentProps {
+  isSelectionModeNone?: boolean;
+}
